Rename school service files
