+++ apps/server/src/modules/school/service/index.ts
@@ -1,4 +1,4 @@
-export * from './school.service';
+export * from './legacy-school.service';
 export * from './school-year.service';
 export * from './federal-state.service';
 export * from './validation';
