+++ apps/server/src/modules/school/service/legacy-school.service.ts
