+++ apps/server/src/modules/video-conference/uc/video-conference-deprecated.uc.ts
@@ -24,7 +24,7 @@ import {
 	AuthorizationContextBuilder,
 	AuthorizationService,
 } from '@src/modules/authorization';
-import { LegacySchoolService } from '@src/modules/school/service/school.service';
+import { LegacySchoolService } from '@src/modules/school';
 import { CourseService } from '@src/modules/learnroom/service/course.service';
 import { UserService } from '@src/modules/user';
 import { IScopeInfo, VideoConference, VideoConferenceInfo, VideoConferenceJoin, VideoConferenceState } from './dto';
